app/backend/requirements.txt
fastapi
uvicorn
boto3
python-dotenv
requests
pydantic[email]
sqlalchemy
psycopg2-binary
